backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mysql-connector-python==8.0.33
httpx==0.27.0
pydantic==2.9.2